Postfix operator (factorial)
There are multiple ways to implement this. For example, in the compiler we could calculate factorial already and just generate the bytecode :putobject with the result.

Of course, factorial here is super inefficient, but it is just an example. The idea is to show how to implement a postfix operator.